requirements.txt
flask
python-dotenv
psycopg2-binary
bs4
requests